requirements.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==3.0.5
channels-redis==3.3.1
charset-normalizer==3.3.0
constantly==15.1.0
cryptography==41.0.4
daphne==3.0.2
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.57.0
django-debug-toolbar==4.2.0
django-etc==1.4.0
django-hitcount==1.3.5
django-phonenumber-field==7.1.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.7
oauthlib==3.2.2
Pillow==10.0.0
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
zope.interface==6.1
